Added 'DIRSEP' constant as directory seprator symbol. Added VIEWPATH constant as `views` directory in path 'app\views'.
Added main view class (to app directory).
Added render as a static method to view class.
render method gets viewPath and checks if it is readable will include it.

Added render method to controllerParent and passed an array to that (for values in the pure view page).

Used render method in 'showAction' in userController.